Better check for invalid Unicode encodings.
Check against overlong encodings for UTF-8, or values that are encoded in
more bytes than required. For example, the value 0 encoded as 0xC0 0x80.
This ensures there is exactly one valid encoding for any given value, which
avoids potential security pitfalls. (e.g. checking for NUL terminator
doesn't match the first encoded 0 value)

Check for invalid Unicode values, which are the values used for UTF-16
surrogate pairs (0xD800-0xDFFF) and values larger than 0x10FFFF. These
checks are performed both during encoding and decoding. Added
dsIsUnicodeCodepointValid() function to allow higher-level code to check
directly.

Unicode functions now operate on size_t. While the text display system will
likely choke well before it exceeds a 32-bit index, the base Unicode
handling shouldn't be limited in this way. This also avoids potential
overflow conditions when processing very long strings on 64-bit systems,
which could lead to infinite loops. With size_t, there should be no
situation where it could overflow as the buffer can't be close enough to
SIZE_MAX where an overflow could occur.

Prefer *_MAX macros over casting -1 for constants.

+++ modules/Text/src/Font.c
@@ -43,7 +43,7 @@
 
 #define INTERSECT_EPSILON 1e-5f
 
-typedef uint32_t (*NextCodepointFunction)(const void* string, uint32_t* index);
+typedef uint32_t (*NextCodepointFunction)(const void* string, size_t* index);
 
 typedef enum PolygonResult
 {
@@ -211,7 +211,7 @@ static bool preloadGlyphs(dsFont* font, dsCommandBuffer* commandBuffer, const vo
 
 	dsFaceGroup_lock(font->group);
 
-	uint32_t index = 0;
+	size_t index = 0;
 	do
 	{
 		uint32_t codepoint = nextCodepointFunc(string, &index);
